Combine schema tests
