--- docs/application-forms/under10k/schema.test.js
@@ -1,22 +1,0 @@
-/* eslint-env jest */
-'use strict';
-const fs = require('fs');
-const path = require('path');
-
-const formBuilder = require('../form');
-
-const form = formBuilder();
-
-const docContents = fs.readFileSync(
-    path.resolve(__dirname, './schema.md'),
-    'utf8'
-);
-
-test.each(Object.keys(form.schema.describe().children))(
-    '%s has documentation entry',
-    (fieldName) => {
-        const fieldNameRegexp = new RegExp(`### ${fieldName}`);
-        const match = fieldNameRegexp.test(docContents);
-        expect(match).toBeTruthy();
-    }
-);
